+++ src/main/java/com/qstpss/mockserver/model/Status.java
@@ -1,4 +1,4 @@
-package com.qstpss.mockserver;
+package com.qstpss.mockserver.model;
 
 public enum Status {
     IN_PROGRESS,
